Use IP addresses in load balancer configuration
Load balancer may run on the same host as the DNS server. This host
may not be configured to itself use the DNS server, though. In this
scenario the load balancer will not be able to connect to the
backend servers as the DNS resolution fails. Using IP addresses
instead of host names in the haproxy.cfg prevents this issue.