Fix virtual machine creation on vSphere 7.0.2
vSphere detects that the ddb.toolsVersion is not set in the VMDK image
and throws an error:

"Customization of the guest operating system is not supported due to the
given reason: Tools is not installed in the GuestOS. Please install the
latest version of open-vm-tools or VMware Tools to enable
GuestCustomization."

The error is triggered by the network configuration. Splitting off
network configuratin into a separate Ansible task seems to fix the
issue.