Add machineNetwork key to install-config
When using a proxy to connect to the internet, MachineNetwork must be properly configured so that OpenShift can automatically add it to NO_PROXY. Otherwise machineNetwork defaults to 10.0.0.0/16 and inter-node traffic will try to use the proxy. This results in the nodes being unable to communicate with each other, and the API server will fail to start.